销售地点管理模块完成

+++ src/views/layout/Header.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="toolbar">
     <span style="font-size: 15px; color: rgb(162, 165, 180);">欢迎来到慧医数字医疗应用系统，当前用户：</span>
-    <span style="font-size: 16px; color: rgb(136, 141, 160); font-style: italic; font-weight: bold; margin-right: 20px;">{{ user }}</span>
+    <span style="font-size: 16px; color: rgb(136, 141, 160); font-style: italic; font-weight: bold; margin-right: 20px;">{{ user == null ? "未登录" : user }}</span>
     <el-button type="danger" round @click="logout">退出登录</el-button>
   </div>
 </template>
@@ -25,8 +25,23 @@ export default {
         type: "warning",
         center: true,
       }).then(() => {
-        this.$router.push("/");
-        sessionStorage.clear();
+        this.$http
+        .post(`/user/logout`)
+        .then((res) => {
+          console.log("用户退出登录res");
+          console.log(res);
+          if (res.data.code == 200) {
+            this.$message({
+              message: this.data.message,
+              type: "success",
+            });
+            this.$router.push("/");
+            sessionStorage.clear();
+          } else {
+            alert(res.data.message);
+          }
+        });
+        
       });
     },
   },
